+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 WordPress-Solutions plugin changes are logged here using <a href="http://semver.org/">Semantic Versioning</a>.
 
+## 2.10.3 (2020-12-04) ##
+* Added remove_directory() method to the USI_WordPress_Solutions_Static class.
+
 ## 2.10.2 (2020-11-18) ##
 * Added eXporter to settings to export posts, updated popup usage list, added money settings field type.